LPD-89588 Document major bump
