+++ modules/apps/oauth2-provider/oauth2-provider-service/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay OAuth2 Provider Service
 Bundle-SymbolicName: com.liferay.oauth2.provider.service
 Bundle-Version: 4.0.112
-Liferay-Require-SchemaVersion: 4.3.0
+Liferay-Require-SchemaVersion: 5.0.0
 Liferay-Service: true
 -dsannotations-options: inherit
