+++ modules/apps/oauth2-provider/oauth2-provider-service/src/main/java/com/liferay/oauth2/provider/internal/upgrade/registry/OAuth2ServiceUpgradeStepRegistrator.java
@@ -151,8 +151,8 @@ public class OAuth2ServiceUpgradeStepRegistrator
 				"OAuth2Authorization", "audiences TEXT null"));
 
 		registry.register(
-			"4.2.8", "4.3.0",
-			new com.liferay.oauth2.provider.internal.upgrade.v4_3_0.
+			"4.2.8", "5.0.0",
+			new com.liferay.oauth2.provider.internal.upgrade.v5_0_0.
 				OAuth2ApplicationIndexedColumnSizeUpgradeProcess());
 	}
