+++ modules/apps/object/object-service/bnd.bnd
@@ -2,6 +2,6 @@ Bundle-Name: Liferay Object Service
 Bundle-SymbolicName: com.liferay.object.service
 Bundle-Version: 1.0.248
 Liferay-Client-Extension-Batch: com/liferay/object/internal/batch
-Liferay-Require-SchemaVersion: 12.2.0
+Liferay-Require-SchemaVersion: 13.0.0
 Liferay-Service: true
 -dsannotations-options: inherit
