+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/upgrade/registry/ObjectServiceUpgradeStepRegistrator.java
@@ -715,8 +715,8 @@ public class ObjectServiceUpgradeStepRegistrator
 				ObjectEntryPicklistDefaultValueUpgradeProcess());
 
 		registry.register(
-			"12.1.1", "12.2.0",
-			new com.liferay.object.internal.upgrade.v12_2_0.
+			"12.1.1", "13.0.0",
+			new com.liferay.object.internal.upgrade.v13_0_0.
 				ObjectEntryIndexedColumnSizeUpgradeProcess());
 	}
