+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/upgrade/v13_0_0/ObjectEntryIndexedColumnSizeUpgradeProcess.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.upgrade.v12_2_0;
+package com.liferay.object.internal.upgrade.v13_0_0;
 
 import com.liferay.portal.kernel.upgrade.BaseIndexedColumnSizeUpgradeProcess;
