+++ modules/apps/object/object-test/src/testIntegration/java/com/liferay/object/internal/upgrade/v13_0_0/test/ObjectEntryIndexedColumnSizeUpgradeProcessTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.upgrade.v12_2_0.test;
+package com.liferay.object.internal.upgrade.v13_0_0.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.string.StringBundler;
